fix(arnes): harden measurement result reads
