+++ tooling/arnes/src/measure/result/finish/tests.rs
@@ -1,5 +1,6 @@
-use super::append_then_commit;
+use super::{append_then_commit, preflight};
 use crate::measure::MeasureError;
+use crate::measure::result::{FinishArgs, MergeReady};
 use std::fs::OpenOptions;
 use std::io::{Read, Seek, SeekFrom};
 
@@ -29,4 +30,29 @@ fn commit_failure_truncates_the_appended_event_to_its_previous_offset() {
     let mut content = String::new();
     file.read_to_string(&mut content).unwrap();
     assert_eq!(content, "{\"event\":\"initial\"}\n");
+}
+
+fn args_with_evidence(size: usize) -> FinishArgs {
+    FinishArgs {
+        run_id: "a".repeat(64),
+        merge_ready: MergeReady::Pass,
+        human_minutes: 1.0,
+        human_edited_diff: false,
+        failure_reason: None,
+        evidence: vec!["x".repeat(size)],
+        regression: false,
+        invariant: Vec::new(),
+    }
+}
+
+#[test]
+fn result_preflight_accepts_a_payload_near_the_reader_limit() {
+    preflight(&args_with_evidence(1_099_000)).unwrap();
+}
+
+#[test]
+fn result_preflight_rejects_an_oversized_payload() {
+    let error = preflight(&args_with_evidence(1_100_000)).unwrap_err();
+
+    assert!(error.to_string().contains("exceeds 1100000 bytes"));
 }
