+++ tooling/arnes/src/measure/result/io.rs
@@ -1,4 +1,5 @@
 use super::super::MeasureError;
+use super::super::store::MAX_RECORD_BYTES;
 use super::super::store::validation::{ensure_regular_file, ensure_single_link};
 use serde::Serialize;
 use serde::de::DeserializeOwned;
@@ -7,8 +8,6 @@ use std::fs::{File, OpenOptions};
 use std::io::{BufRead, BufReader, Read, Seek, SeekFrom};
 use std::os::unix::fs::OpenOptionsExt;
 use std::path::Path;
-
-const MAX_RECORD_BYTES: usize = 1_100_000;
 
 pub fn read_optional_json<T>(path: &Path, label: &str) -> Result<Option<T>, MeasureError>
 where
@@ -45,56 +44,47 @@ where
     Ok(record)
 }
 
-pub fn read_jsonl_typed<T>(path: &Path, label: &str) -> Result<Vec<T>, MeasureError>
+pub fn visit_jsonl_typed<T>(
+    path: &Path,
+    label: &str,
+    visitor: impl FnMut(T) -> Result<(), MeasureError>,
+) -> Result<(), MeasureError>
 where
     T: DeserializeOwned + Serialize,
 {
-    parse_typed(read_lines(path, label)?, label)
-}
-
-pub fn read_jsonl_typed_file<T>(file: &mut File, label: &str) -> Result<Vec<T>, MeasureError>
-where
-    T: DeserializeOwned + Serialize,
-{
-    parse_typed(read_lines_file(file, label)?, label)
-}
-
-fn parse_typed<T>(lines: Vec<Vec<u8>>, label: &str) -> Result<Vec<T>, MeasureError>
-where
-    T: DeserializeOwned + Serialize,
-{
-    lines
-        .into_iter()
-        .map(|line| {
-            let record: T = serde_json::from_slice(&line)
-                .map_err(|_| MeasureError::new(format!("managed {label} has an invalid record")))?;
-            let raw: Value = serde_json::from_slice(&line)?;
-            if serde_json::to_value(&record)? != raw {
-                return Err(MeasureError::new(format!(
-                    "managed {label} has an invalid record"
-                )));
-            }
-            Ok(record)
-        })
-        .collect()
-}
-
-fn read_lines(path: &Path, label: &str) -> Result<Vec<Vec<u8>>, MeasureError> {
     match std::fs::symlink_metadata(path) {
         Ok(_) => {}
-        Err(error) if error.kind() == std::io::ErrorKind::NotFound => return Ok(Vec::new()),
+        Err(error) if error.kind() == std::io::ErrorKind::NotFound => return Ok(()),
         Err(error) => return Err(error.into()),
     }
     ensure_regular_file(path)?;
     let mut file = open_read(path)?;
     file.lock()?;
-    read_lines_file(&mut file, label)
+    visit_jsonl_typed_file(&mut file, label, visitor)
 }
 
-fn read_lines_file(file: &mut File, label: &str) -> Result<Vec<Vec<u8>>, MeasureError> {
+pub fn open_locked_jsonl(path: &Path) -> Result<Option<File>, MeasureError> {
+    match std::fs::symlink_metadata(path) {
+        Ok(_) => {}
+        Err(error) if error.kind() == std::io::ErrorKind::NotFound => return Ok(None),
+        Err(error) => return Err(error.into()),
+    }
+    ensure_regular_file(path)?;
+    let file = open_read(path)?;
+    file.lock()?;
+    Ok(Some(file))
+}
+
+pub fn visit_jsonl_typed_file<T>(
+    file: &mut File,
+    label: &str,
+    mut visitor: impl FnMut(T) -> Result<(), MeasureError>,
+) -> Result<(), MeasureError>
+where
+    T: DeserializeOwned + Serialize,
+{
     file.seek(SeekFrom::Start(0))?;
     let mut reader = BufReader::new(file);
-    let mut records = Vec::new();
     loop {
         let mut line = Vec::new();
         let read = reader
@@ -110,9 +100,24 @@ fn read_lines_file(file: &mut File, label: &str) -> Result<Vec<Vec<u8>>, Measure
             )));
         }
         line.pop();
-        records.push(line);
+        visitor(parse_typed(&line, label)?)?;
     }
-    Ok(records)
+    Ok(())
+}
+
+fn parse_typed<T>(line: &[u8], label: &str) -> Result<T, MeasureError>
+where
+    T: DeserializeOwned + Serialize,
+{
+    let record: T = serde_json::from_slice(line)
+        .map_err(|_| MeasureError::new(format!("managed {label} has an invalid record")))?;
+    let raw: Value = serde_json::from_slice(line)?;
+    if serde_json::to_value(&record)? != raw {
+        return Err(MeasureError::new(format!(
+            "managed {label} has an invalid record"
+        )));
+    }
+    Ok(record)
 }
 
 fn open_read(path: &Path) -> Result<File, MeasureError> {
@@ -122,4 +127,36 @@ fn open_read(path: &Path) -> Result<File, MeasureError> {
         .open(path)?;
     ensure_single_link(&file.metadata()?, path)?;
     Ok(file)
+}
+
+#[cfg(test)]
+mod tests {
+    use super::visit_jsonl_typed;
+    use serde::{Deserialize, Serialize};
+    use std::io::Write;
+
+    #[derive(Deserialize, Serialize)]
+    struct Record {
+        sequence: usize,
+    }
+
+    #[test]
+    fn visits_a_long_journal_line_by_line() {
+        let mut file = tempfile::NamedTempFile::new().unwrap();
+        for sequence in 0..20_000 {
+            writeln!(file, "{{\"sequence\":{sequence}}}").unwrap();
+        }
+        let mut count = 0;
+        let mut last = None;
+
+        visit_jsonl_typed::<Record>(file.path(), "records.jsonl", |record| {
+            count += 1;
+            last = Some(record.sequence);
+            Ok(())
+        })
+        .unwrap();
+
+        assert_eq!(count, 20_000);
+        assert_eq!(last, Some(19_999));
+    }
 }
